Bug 1289156 - Final tweaks for libmysqlclient deploy on Heroku/SCL3
There's a strange bug when creating relative symlinks on top of an
already existing symlink, in that the target resolves to the wrong file.
As such, we explicitly check for an existing symlink first, to avoid
overwriting.

The SCL3 system check disabling has been tweaked, since the `HOSTNAME`
environment variable doesn't appear to be set during deploy, even though
it appears in `env` on the same machine.

Yey for deployment code that can only really be tested by deploying \o/